Completed more functions, such as energy and force calculations as well as total integration function. Only need to write velocityVerlet() method. Python placeholder code in there now. Needs to be changed to C++

+++ main.cpp
@@ -3,7 +3,7 @@
 
 int main() {
     // Simulation variables
-    int num_iter = 2000;
+    long num_iter = 2000;
     int num_simulations = 1;
     unsigned long n_particle = 256;
     double temp = 1.0;
